--- springboot/trading/src/main/java/ca/jrvs/apps/trading/account/AccountDao.java
@@ -1,7 +1,0 @@
-package ca.jrvs.apps.trading.account;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AccountDao extends JpaRepository<Account, Integer> {
-
-}
