+++ springboot/trading/src/main/java/ca/jrvs/apps/trading/dto/IexQuote.java
@@ -1,0 +1,754 @@
+package ca.jrvs.apps.trading.dto;
+
+import java.io.Serializable;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+
+// @JsonInclude(JsonInclude.Include.NON_NULL)
+@JsonPropertyOrder({
+        "avgTotalVolume",
+        "calculationPrice",
+        "change",
+        "changePercent",
+        "close",
+        "closeSource",
+        "closeTime",
+        "companyName",
+        "currency",
+        "delayedPrice",
+        "delayedPriceTime",
+        "extendedChange",
+        "extendedChangePercent",
+        "extendedPrice",
+        "extendedPriceTime",
+        "high",
+        "highSource",
+        "highTime",
+        "iexAskPrice",
+        "iexAskSize",
+        "iexBidPrice",
+        "iexBidSize",
+        "iexClose",
+        "iexCloseTime",
+        "iexLastUpdated",
+        "iexMarketPercent",
+        "iexOpen",
+        "iexOpenTime",
+        "iexRealtimePrice",
+        "iexRealtimeSize",
+        "iexVolume",
+        "lastTradeTime",
+        "latestPrice",
+        "latestSource",
+        "latestTime",
+        "latestUpdate",
+        "latestVolume",
+        "low",
+        "lowSource",
+        "lowTime",
+        "marketCap",
+        "oddLotDelayedPrice",
+        "oddLotDelayedPriceTime",
+        "open",
+        "openTime",
+        "openSource",
+        "peRatio",
+        "previousClose",
+        "previousVolume",
+        "primaryExchange",
+        "symbol",
+        "volume",
+        "week52High",
+        "week52Low",
+        "ytdChange",
+        "isUSMarketOpen"
+})
+
+public class IexQuote implements Serializable {
+
+    @JsonProperty("avgTotalVolume")
+    private Long avgTotalVolume;
+    @JsonProperty("calculationPrice")
+    private String calculationPrice;
+    @JsonProperty("change")
+    private Double change;
+    @JsonProperty("changePercent")
+    private Double changePercent;
+    @JsonProperty("close")
+    private Double close;
+    @JsonProperty("closeSource")
+    private String closeSource;
+    @JsonProperty("closeTime")
+    private Long closeTime;
+    @JsonProperty("companyName")
+    private String companyName;
+    @JsonProperty("currency")
+    private String currency;
+    @JsonProperty("delayedPrice")
+    private Double delayedPrice;
+    @JsonProperty("delayedPriceTime")
+    private Long delayedPriceTime;
+    @JsonProperty("extendedChange")
+    private Double extendedChange;
+    @JsonProperty("extendedChangePercent")
+    private Double extendedChangePercent;
+    @JsonProperty("extendedPrice")
+    private Double extendedPrice;
+    @JsonProperty("extendedPriceTime")
+    private Long extendedPriceTime;
+    @JsonProperty("high")
+    private Double high;
+    @JsonProperty("highSource")
+    private String highSource;
+    @JsonProperty("highTime")
+    private Long highTime;
+    @JsonProperty("iexAskPrice")
+    private Double iexAskPrice;
+    @JsonProperty("iexAskSize")
+    private Long iexAskSize;
+    @JsonProperty("iexBidPrice")
+    private Double iexBidPrice;
+    @JsonProperty("iexBidSize")
+    private Long iexBidSize;
+    @JsonProperty("iexClose")
+    private Double iexClose;
+    @JsonProperty("iexCloseTime")
+    private Long iexCloseTime;
+    @JsonProperty("iexLastUpdated")
+    private Long iexLastUpdated;
+    @JsonProperty("iexMarketPercent")
+    private Double iexMarketPercent;
+    @JsonProperty("iexOpen")
+    private Double iexOpen;
+    @JsonProperty("iexOpenTime")
+    private Long iexOpenTime;
+    @JsonProperty("iexRealtimePrice")
+    private Double iexRealtimePrice;
+    @JsonProperty("iexRealtimeSize")
+    private Long iexRealtimeSize;
+    @JsonProperty("iexVolume")
+    private Long iexVolume;
+    @JsonProperty("lastTradeTime")
+    private Long lastTradeTime;
+    @JsonProperty("latestPrice")
+    private Double latestPrice;
+    @JsonProperty("latestSource")
+    private String latestSource;
+    @JsonProperty("latestTime")
+    private String latestTime;
+    @JsonProperty("latestUpdate")
+    private Long latestUpdate;
+    @JsonProperty("latestVolume")
+    private Long latestVolume;
+    @JsonProperty("low")
+    private Double low;
+    @JsonProperty("lowSource")
+    private String lowSource;
+    @JsonProperty("lowTime")
+    private Double lowTime;
+    @JsonProperty("marketCap")
+    private Long marketCap;
+    @JsonProperty("oddLotDelayedPrice")
+    private Double oddLotDelayedPrice;
+    @JsonProperty("oddLotDelayedPriceTime")
+    private Long oddLotDelayedPriceTime;
+    @JsonProperty("open")
+    private Double open;
+    @JsonProperty("openTime")
+    private Long openTime;
+    @JsonProperty("openSource")
+    private String openSource;
+    @JsonProperty("peRatio")
+    private Double peRatio;
+    @JsonProperty("previousClose")
+    private Double previousClose;
+    @JsonProperty("previousVolume")
+    private Long previousVolume;
+    @JsonProperty("primaryExchange")
+    private String primaryExchange;
+    @JsonProperty("symbol")
+    private String symbol;
+    @JsonProperty("volume")
+    private Long volume;
+    @JsonProperty("week52High")
+    private Double week52High;
+    @JsonProperty("week52Low")
+    private Double week52Low;
+    @JsonProperty("ytdChange")
+    private Double ytdChange;
+    @JsonProperty("isUSMarketOpen")
+    private Boolean isUSMarketOpen;
+//    @JsonIgnore
+//    private final Map<String, Object> additionalProperties = new LinkedHashMap<String, Object>();
+
+    @JsonProperty("avgTotalVolume")
+    public Long getAvgTotalVolume() {
+        return avgTotalVolume;
+    }
+
+    @JsonProperty("avgTotalVolume")
+    public void setAvgTotalVolume(Long avgTotalVolume) {
+        this.avgTotalVolume = avgTotalVolume;
+    }
+
+    @JsonProperty("calculationPrice")
+    public String getCalculationPrice() {
+        return calculationPrice;
+    }
+
+    @JsonProperty("calculationPrice")
+    public void setCalculationPrice(String calculationPrice) {
+        this.calculationPrice = calculationPrice;
+    }
+
+    @JsonProperty("change")
+    public Double getChange() {
+        return change;
+    }
+
+    @JsonProperty("change")
+    public void setChange(Double change) {
+        this.change = change;
+    }
+
+    @JsonProperty("changePercent")
+    public Double getChangePercent() {
+        return changePercent;
+    }
+
+    @JsonProperty("changePercent")
+    public void setChangePercent(Double changePercent) {
+        this.changePercent = changePercent;
+    }
+
+    @JsonProperty("close")
+    public Double getClose() {
+        return close;
+    }
+
+    @JsonProperty("close")
+    public void setClose(Double close) {
+        this.close = close;
+    }
+
+    @JsonProperty("closeSource")
+    public String getCloseSource() {
+        return closeSource;
+    }
+
+    @JsonProperty("closeSource")
+    public void setCloseSource(String closeSource) {
+        this.closeSource = closeSource;
+    }
+
+    @JsonProperty("closeTime")
+    public Long getCloseTime() {
+        return closeTime;
+    }
+
+    @JsonProperty("closeTime")
+    public void setCloseTime(Long closeTime) {
+        this.closeTime = closeTime;
+    }
+
+    @JsonProperty("companyName")
+    public String getCompanyName() {
+        return companyName;
+    }
+
+    @JsonProperty("companyName")
+    public void setCompanyName(String companyName) {
+        this.companyName = companyName;
+    }
+
+    @JsonProperty("currency")
+    public String getCurrency() {
+        return currency;
+    }
+
+    @JsonProperty("currency")
+    public void setCurrency(String currency) {
+        this.currency = currency;
+    }
+
+    @JsonProperty("delayedPrice")
+    public Double getDelayedPrice() {
+        return delayedPrice;
+    }
+
+    @JsonProperty("delayedPrice")
+    public void setDelayedPrice(Double delayedPrice) {
+        this.delayedPrice = delayedPrice;
+    }
+
+    @JsonProperty("delayedPriceTime")
+    public Long getDelayedPriceTime() {
+        return delayedPriceTime;
+    }
+
+    @JsonProperty("delayedPriceTime")
+    public void setDelayedPriceTime(Long delayedPriceTime) {
+        this.delayedPriceTime = delayedPriceTime;
+    }
+
+    @JsonProperty("extendedChange")
+    public Double getExtendedChange() {
+        return extendedChange;
+    }
+
+    @JsonProperty("extendedChange")
+    public void setExtendedChange(Double extendedChange) {
+        this.extendedChange = extendedChange;
+    }
+
+    @JsonProperty("extendedChangePercent")
+    public Double getExtendedChangePercent() {
+        return extendedChangePercent;
+    }
+
+    @JsonProperty("extendedChangePercent")
+    public void setExtendedChangePercent(Double extendedChangePercent) {
+        this.extendedChangePercent = extendedChangePercent;
+    }
+
+    @JsonProperty("extendedPrice")
+    public Double getExtendedPrice() {
+        return extendedPrice;
+    }
+
+    @JsonProperty("extendedPrice")
+    public void setExtendedPrice(Double extendedPrice) {
+        this.extendedPrice = extendedPrice;
+    }
+
+    @JsonProperty("extendedPriceTime")
+    public Long getExtendedPriceTime() {
+        return extendedPriceTime;
+    }
+
+    @JsonProperty("extendedPriceTime")
+    public void setExtendedPriceTime(Long extendedPriceTime) {
+        this.extendedPriceTime = extendedPriceTime;
+    }
+
+    @JsonProperty("high")
+    public Double getHigh() {
+        return high;
+    }
+
+    @JsonProperty("high")
+    public void setHigh(Double high) {
+        this.high = high;
+    }
+
+    @JsonProperty("highSource")
+    public String getHighSource() {
+        return highSource;
+    }
+
+    @JsonProperty("highSource")
+    public void setHighSource(String highSource) {
+        this.highSource = highSource;
+    }
+
+    @JsonProperty("highTime")
+    public Long getHighTime() {
+        return highTime;
+    }
+
+    @JsonProperty("highTime")
+    public void setHighTime(Long highTime) {
+        this.highTime = highTime;
+    }
+
+    @JsonProperty("iexAskPrice")
+    public Double getIexAskPrice() {
+        return iexAskPrice;
+    }
+
+    @JsonProperty("iexAskPrice")
+    public void setIexAskPrice(Double iexAskPrice) {
+        this.iexAskPrice = iexAskPrice;
+    }
+
+    @JsonProperty("iexAskSize")
+    public Long getIexAskSize() {
+        return iexAskSize;
+    }
+
+    @JsonProperty("iexAskSize")
+    public void setIexAskSize(Long iexAskSize) {
+        this.iexAskSize = iexAskSize;
+    }
+
+    @JsonProperty("iexBidPrice")
+    public Double getIexBidPrice() {
+        return iexBidPrice;
+    }
+
+    @JsonProperty("iexBidPrice")
+    public void setIexBidPrice(Double iexBidPrice) {
+        this.iexBidPrice = iexBidPrice;
+    }
+
+    @JsonProperty("iexBidSize")
+    public Long getIexBidSize() {
+        return iexBidSize;
+    }
+
+    @JsonProperty("iexBidSize")
+    public void setIexBidSize(Long iexBidSize) {
+        this.iexBidSize = iexBidSize;
+    }
+
+    @JsonProperty("iexClose")
+    public Double getIexClose() {
+        return iexClose;
+    }
+
+    @JsonProperty("iexClose")
+    public void setIexClose(Double iexClose) {
+        this.iexClose = iexClose;
+    }
+
+    @JsonProperty("iexCloseTime")
+    public Long getIexCloseTime() {
+        return iexCloseTime;
+    }
+
+    @JsonProperty("iexCloseTime")
+    public void setIexCloseTime(Long iexCloseTime) {
+        this.iexCloseTime = iexCloseTime;
+    }
+
+    @JsonProperty("iexLastUpdated")
+    public Long getIexLastUpdated() {
+        return iexLastUpdated;
+    }
+
+    @JsonProperty("iexLastUpdated")
+    public void setIexLastUpdated(Long iexLastUpdated) {
+        this.iexLastUpdated = iexLastUpdated;
+    }
+
+    @JsonProperty("iexMarketPercent")
+    public Double getIexMarketPercent() {
+        return iexMarketPercent;
+    }
+
+    @JsonProperty("iexMarketPercent")
+    public void setIexMarketPercent(Double iexMarketPercent) {
+        this.iexMarketPercent = iexMarketPercent;
+    }
+
+    @JsonProperty("iexOpen")
+    public Double getIexOpen() {
+        return iexOpen;
+    }
+
+    @JsonProperty("iexOpen")
+    public void setIexOpen(Double iexOpen) {
+        this.iexOpen = iexOpen;
+    }
+
+    @JsonProperty("iexOpenTime")
+    public Long getIexOpenTime() {
+        return iexOpenTime;
+    }
+
+    @JsonProperty("iexOpenTime")
+    public void setIexOpenTime(Long iexOpenTime) {
+        this.iexOpenTime = iexOpenTime;
+    }
+
+    @JsonProperty("iexRealtimePrice")
+    public Double getIexRealtimePrice() {
+        return iexRealtimePrice;
+    }
+
+    @JsonProperty("iexRealtimePrice")
+    public void setIexRealtimePrice(Double iexRealtimePrice) {
+        this.iexRealtimePrice = iexRealtimePrice;
+    }
+
+    @JsonProperty("iexRealtimeSize")
+    public Long getIexRealtimeSize() {
+        return iexRealtimeSize;
+    }
+
+    @JsonProperty("iexRealtimeSize")
+    public void setIexRealtimeSize(Long iexRealtimeSize) {
+        this.iexRealtimeSize = iexRealtimeSize;
+    }
+
+    @JsonProperty("iexVolume")
+    public Long getIexVolume() {
+        return iexVolume;
+    }
+
+    @JsonProperty("iexVolume")
+    public void setIexVolume(Long iexVolume) {
+        this.iexVolume = iexVolume;
+    }
+
+    @JsonProperty("lastTradeTime")
+    public Long getLastTradeTime() {
+        return lastTradeTime;
+    }
+
+    @JsonProperty("lastTradeTime")
+    public void setLastTradeTime(Long lastTradeTime) {
+        this.lastTradeTime = lastTradeTime;
+    }
+
+    @JsonProperty("latestPrice")
+    public Double getLatestPrice() {
+        return latestPrice;
+    }
+
+    @JsonProperty("latestPrice")
+    public void setLatestPrice(Double latestPrice) {
+        this.latestPrice = latestPrice;
+    }
+
+    @JsonProperty("latestSource")
+    public String getLatestSource() {
+        return latestSource;
+    }
+
+    @JsonProperty("latestSource")
+    public void setLatestSource(String latestSource) {
+        this.latestSource = latestSource;
+    }
+
+    @JsonProperty("latestTime")
+    public String getLatestTime() {
+        return latestTime;
+    }
+
+    @JsonProperty("latestTime")
+    public void setLatestTime(String latestTime) {
+        this.latestTime = latestTime;
+    }
+
+    @JsonProperty("latestUpdate")
+    public Long getLatestUpdate() {
+        return latestUpdate;
+    }
+
+    @JsonProperty("latestUpdate")
+    public void setLatestUpdate(Long latestUpdate) {
+        this.latestUpdate = latestUpdate;
+    }
+
+    @JsonProperty("latestVolume")
+    public Long getLatestVolume() {
+        return latestVolume;
+    }
+
+    @JsonProperty("latestVolume")
+    public void setLatestVolume(Long latestVolume) {
+        this.latestVolume = latestVolume;
+    }
+
+    @JsonProperty("low")
+    public Double getLow() {
+        return low;
+    }
+
+    @JsonProperty("low")
+    public void setLow(Double low) {
+        this.low = low;
+    }
+
+    @JsonProperty("lowSource")
+    public String getLowSource() {
+        return lowSource;
+    }
+
+    @JsonProperty("lowSource")
+    public void setLowSource(String lowSource) {
+        this.lowSource = lowSource;
+    }
+
+    @JsonProperty("lowTime")
+    public Double getLowTime() {
+        return lowTime;
+    }
+
+    @JsonProperty("lowTime")
+    public void setLowTime(Double lowTime) {
+        this.lowTime = lowTime;
+    }
+
+    @JsonProperty("marketCap")
+    public Long getMarketCap() {
+        return marketCap;
+    }
+
+    @JsonProperty("marketCap")
+    public void setMarketCap(Long marketCap) {
+        this.marketCap = marketCap;
+    }
+
+    @JsonProperty("oddLotDelayedPrice")
+    public Double getOddLotDelayedPrice() {
+        return oddLotDelayedPrice;
+    }
+
+    @JsonProperty("oddLotDelayedPrice")
+    public void setOddLotDelayedPrice(Double oddLotDelayedPrice) {
+        this.oddLotDelayedPrice = oddLotDelayedPrice;
+    }
+
+    @JsonProperty("oddLotDelayedPriceTime")
+    public Long getOddLotDelayedPriceTime() {
+        return oddLotDelayedPriceTime;
+    }
+
+    @JsonProperty("oddLotDelayedPriceTime")
+    public void setOddLotDelayedPriceTime(Long oddLotDelayedPriceTime) {
+        this.oddLotDelayedPriceTime = oddLotDelayedPriceTime;
+    }
+
+    @JsonProperty("open")
+    public Double getOpen() {
+        return open;
+    }
+
+    @JsonProperty("open")
+    public void setOpen(Double open) {
+        this.open = open;
+    }
+
+    @JsonProperty("openTime")
+    public Long getOpenTime() {
+        return openTime;
+    }
+
+    @JsonProperty("openTime")
+    public void setOpenTime(Long openTime) {
+        this.openTime = openTime;
+    }
+
+    @JsonProperty("openSource")
+    public String getOpenSource() {
+        return openSource;
+    }
+
+    @JsonProperty("openSource")
+    public void setOpenSource(String openSource) {
+        this.openSource = openSource;
+    }
+
+    @JsonProperty("peRatio")
+    public Double getPeRatio() {
+        return peRatio;
+    }
+
+    @JsonProperty("peRatio")
+    public void setPeRatio(Double peRatio) {
+        this.peRatio = peRatio;
+    }
+
+    @JsonProperty("previousClose")
+    public Double getPreviousClose() {
+        return previousClose;
+    }
+
+    @JsonProperty("previousClose")
+    public void setPreviousClose(Double previousClose) {
+        this.previousClose = previousClose;
+    }
+
+    @JsonProperty("previousVolume")
+    public Long getPreviousVolume() {
+        return previousVolume;
+    }
+
+    @JsonProperty("previousVolume")
+    public void setPreviousVolume(Long previousVolume) {
+        this.previousVolume = previousVolume;
+    }
+
+    @JsonProperty("primaryExchange")
+    public String getPrimaryExchange() {
+        return primaryExchange;
+    }
+
+    @JsonProperty("primaryExchange")
+    public void setPrimaryExchange(String primaryExchange) {
+        this.primaryExchange = primaryExchange;
+    }
+
+    @JsonProperty("symbol")
+    public String getSymbol() {
+        return symbol;
+    }
+
+    @JsonProperty("symbol")
+    public void setSymbol(String symbol) {
+        this.symbol = symbol;
+    }
+
+    @JsonProperty("volume")
+    public Long getVolume() {
+        return volume;
+    }
+
+    @JsonProperty("volume")
+    public void setVolume(Long volume) {
+        this.volume = volume;
+    }
+
+    @JsonProperty("week52High")
+    public Double getWeek52High() {
+        return week52High;
+    }
+
+    @JsonProperty("week52High")
+    public void setWeek52High(Double week52High) {
+        this.week52High = week52High;
+    }
+
+    @JsonProperty("week52Low")
+    public Double getWeek52Low() {
+        return week52Low;
+    }
+
+    @JsonProperty("week52Low")
+    public void setWeek52Low(Double week52Low) {
+        this.week52Low = week52Low;
+    }
+
+    @JsonProperty("ytdChange")
+    public Double getYtdChange() {
+        return ytdChange;
+    }
+
+    @JsonProperty("ytdChange")
+    public void setYtdChange(Double ytdChange) {
+        this.ytdChange = ytdChange;
+    }
+
+    @JsonProperty("isUSMarketOpen")
+    public Boolean isIsUSMarketOpen() {
+        return isUSMarketOpen;
+    }
+
+    @JsonProperty("isUSMarketOpen")
+    public void setIsUSMarketOpen(Boolean isUSMarketOpen) {
+        this.isUSMarketOpen = isUSMarketOpen;
+    }
+
+//    @JsonAnyGetter
+//    public Map<String, Object> getAdditionalProperties() {
+//        return this.additionalProperties;
+//    }
+//
+//    @JsonAnySetter
+//    public void setAdditionalProperty(String name, Object value) {
+//        this.additionalProperties.put(name, value);
+//    }
+}
