--- springboot/trading/src/main/java/ca/jrvs/apps/trading/position/PositionDao.java
@@ -1,6 +1,0 @@
-package ca.jrvs.apps.trading.position;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PositionDao extends JpaRepository<Position, Integer> {
-}
