--- springboot/trading/src/main/java/ca/jrvs/apps/trading/quote/QuoteDao.java
@@ -1,9 +1,0 @@
-package ca.jrvs.apps.trading.quote;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface QuoteDao extends JpaRepository<Quote, String> {
-
-}
