+++ springboot/trading/src/main/java/ca/jrvs/apps/trading/repository/PositionDao.java
@@ -1,0 +1,7 @@
+package ca.jrvs.apps.trading.repository;
+
+import ca.jrvs.apps.trading.domain.Position;
+
+public interface PositionDao extends ReadOnlyRepository<Position, Integer> {
+
+}
