+++ springboot/trading/src/main/java/ca/jrvs/apps/trading/repository/QuoteDao.java
@@ -1,0 +1,10 @@
+package ca.jrvs.apps.trading.repository;
+
+import ca.jrvs.apps.trading.domain.Quote;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface QuoteDao extends JpaRepository<Quote, String> {
+
+}
