--- springboot/trading/src/main/java/ca/jrvs/apps/trading/securityorder/SecurityOrderDao.java
@@ -1,7 +1,0 @@
-package ca.jrvs.apps.trading.securityorder;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SecurityOrderDao extends JpaRepository<SecurityOrder, Integer> {
-
-}
