+++ springboot/trading/src/main/java/ca/jrvs/apps/trading/service/TraderAccountService.java
@@ -1,0 +1,5 @@
+package ca.jrvs.apps.trading.service;
+
+public class TraderAccountService {
+
+}
