--- springboot/trading/src/main/java/ca/jrvs/apps/trading/trader/Trader.java
@@ -1,68 +1,0 @@
-package ca.jrvs.apps.trading.trader;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-
-import java.sql.Date;
-
-@Entity
-public class Trader {
-
-    // Primary Key
-    @Id
-    private Integer id;
-
-    private String first_name;
-    private String last_name;
-    private Date dob;
-    private String country;
-    private String email;
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getFirst_name() {
-        return first_name;
-    }
-
-    public void setFirst_name(String first_name) {
-        this.first_name = first_name;
-    }
-
-    public String getLast_name() {
-        return last_name;
-    }
-
-    public void setLast_name(String last_name) {
-        this.last_name = last_name;
-    }
-
-    public Date getDob() {
-        return dob;
-    }
-
-    public void setDob(Date dob) {
-        this.dob = dob;
-    }
-
-    public String getCountry() {
-        return country;
-    }
-
-    public void setCountry(String country) {
-        this.country = country;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-}
