--- springboot/trading/src/main/java/ca/jrvs/apps/trading/trader/TraderDao.java
@@ -1,7 +1,0 @@
-package ca.jrvs.apps.trading.trader;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TraderDao extends JpaRepository<Trader, Integer> {
-
-}
